apps/backend/pyproject.toml
[project]
name = "toast"
version = "0.1.0"
description = "Toast AI"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "aiohttp>=3.11.18",
    "bs4>=0.0.2",
    "certifi>=2025.4.26",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "langchain>=0.3.25",
    "litellm>=1.65.1",
    "markdown>=3.8",
    "markdownify>=1.1.0",
    "motor[zstd]>=3.7.0",
    "mypy>=1.15.0",
    "nanoid>=2.0.0",
    "nltk>=3.9.1",
    "pdfplumber>=0.11.7",
    "pinecone>=6.0.2",
    "psutil>=7.0.0",
    "pydantic>=2.10.6",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.5.0",
    "shortuuid>=1.0.13",
    "streamlit>=1.43.2",
    "streamlit-tags>=1.2.8",
    "structlog>=25.4.0",
    "supabase>=2.13.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

ignore = ["E501", "B008"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.params.Depends",
    "fastapi.Query",
    "fastapi.params.Query",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "E712",
] # Allow explicit boolean comparisons in tests for clarity

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

exclude = ["^\\.venv/", "^build/", "^dist/"]
follow_imports = "skip"

# Ignore specific error types globally
disable_error_code = ["misc"]

[[tool.mypy.overrides]]
module = [
    "cryptography.*",
    "pymongo.*",
    "motor.*",
    "supabase.*",
    "httpx.*",
    "httpcore.*",
]
ignore_errors = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
follow_imports = "skip"
